requirements.txt
Django
django-filter
django-widget-tweaks
djangorestframework
Pillow
xlwt
python-dotenv
psycopg2-binary